backend/requirements.txt
# Feel free to add your Python dependencies below. For formatting guidelines, see:
# https://pip.pypa.io/en/latest/reference/requirements-file-format/

apify >= 2.3.0,< 3
apify-client < 2.0  # Optional for Apify integration
llama-index < 1.0.0
polars < 2.0.0
playwright>=1.40.0
pydantic
typer
pandas
PyYAML
loguru
# --- NEW: for production queue/REST/batching ---
fastapi
uvicorn
redis
requests
python-jose[cryptography]
python-multipart
